+++ tests/SchemaParserTest.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Nwidart\Modules\tests;
+namespace Llama\Modules\tests;
 
-use Nwidart\Modules\Support\Migrations\SchemaParser;
+use Llama\Modules\Support\Migrations\SchemaParser;
 
 class SchemaParserTest extends \PHPUnit_Framework_TestCase
 {